backend/requirements.txt
attrdict==2.0.0
certifi==2018.1.18
chardet==3.0.4
cytoolz==0.9.0
Django==2.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
eth-abi==1.0.0b0
eth-account==0.1.0a2
eth-keyfile==0.5.1
eth-keys==0.2.0b1
eth-rlp==0.1.0a2
eth-utils==1.0.0b1
hexbytes==0.1.0b0
idna==2.6
lru-dict==1.1.6
Markdown==2.6.11
pycryptodome==3.4.7
pysha3==1.0.2
pytz==2018.3
requests==2.18.4
rlp==0.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==4.0.0b9
